cmd/darepocli: extract commands into importable package
Move all command implementations from package main into a new
darepoclicommands sub-package with an exported NewRootCmd()
constructor. This allows other packages (e.g., darepotest repl)
to import and embed the darepocli command tree without compiling
a separate binary.

The main.go becomes a thin wrapper that calls
darepoclicommands.NewRootCmd().Execute().
