+++ .gitignore
@@ -1,5 +1,7 @@
 /arkcli
 /arkd
+/darepocli
+/darepod
 **/test-artifacts
 merge-sql-schemas
