+++ cmd/darepocli/darepoclicommands/client.go
@@ -1,4 +1,4 @@
-package main
+package darepoclicommands
 
 import (
 	"crypto/tls"
