+++ cmd/darepocli/darepoclicommands/cmd_send.go
@@ -1,4 +1,4 @@
-package main
+package darepoclicommands
 
 import (
 	"context"
