+++ cmd/darepocli/darepoclicommands/cmd_wallet.go
@@ -1,4 +1,4 @@
-package main
+package darepoclicommands
 
 import (
 	"bufio"
