+++ cmd/darepocli/darepoclicommands/format.go
@@ -1,4 +1,4 @@
-package main
+package darepoclicommands
 
 import (
 	"encoding/json"
