+++ cmd/darepocli/darepoclicommands/json_input.go
@@ -1,4 +1,4 @@
-package main
+package darepoclicommands
 
 import (
 	"fmt"
